# api/src/openapi.yaml
openapi: 3.1.0
info:
  title: Deffavia API
  version: "1.1.0"
  description: API para reservas de asientos (orden + ítems), exportación/importación XML, reportes y correo.
servers:
  - url: http://localhost:3001
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: alumno@gmail.com }
        password: { type: string, minLength: 6, example: Secret123 }
    RegisterResponse:
      type: object
      properties:
        ok: { type: boolean }
        user:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            created_at: { type: string, format: date-time }
...
    UpdateUserPatchRequest:
      type: object
      properties:
        email: { type: string, format: email, description: "Debe ser @gmail.com u @outlook.com" }
        password: { type: string, minLength: 6 }
        currentPassword: { type: string, description: "Requerida si se cambia password" }
    UpdateUserPatchResponse:
      type: object
      properties:
        ok: { type: boolean }
        user:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            created_at: { type: string, format: date-time }
        updated:
          type: object
          properties:
            email: { type: boolean }
            password: { type: boolean }
paths:
...
  /api/users:
    post:
      tags: [Users]
      summary: Crear usuario (@gmail/@outlook)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '200':
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterResponse' }
        '4XX':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/users/{id}:
    patch:
      tags: [Users]
      summary: Actualizar perfil de usuario (parcial)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserPatchRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UpdateUserPatchResponse' }
        '4XX':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...